+++ src/app/@core/Model/ProjectDetailsAudit.ts
@@ -1,0 +1,6 @@
+export class ProjectDetailsAudit {
+  createdVulnerabilities: number;
+  resolvedVulnerabilities: number;
+  averageTimeToResolve: String;
+  percentCriticalsSolved: String;
+}