+++ src/app/pages/extra-components/details-component.ts
@@ -4,6 +4,16 @@ import {NbWindowService} from '@nebular/theme';
 import {ShowProjectService} from '../../@core/service/ShowProjectService';
 import {Vulnerability} from '../../@core/Model/Vulnerability';
 import {Toast} from '../../@core/utils/Toast';
+import {ProjectAudit} from '../../@core/Model/ProjectAudit';
+import {VulnerabilityHistoryRequest} from '../../@core/Model/VulnerabilityHistoryRequest';
+
+interface EventItem {
+  status?: String;
+  date?: String;
+  icon?: String;
+  color?: String;
+  image?: String;
+}
 
 @Component({
   template: `
@@ -69,6 +79,28 @@ can be marked by Mixeway Vuln Auditor or manualy by a user, read docs to get mor
                 </nb-card>
               </div>
             </div>
+            <div class="row" *ngIf="events">
+              <div class="col-md-12" >
+                <nb-card>
+                  <nb-card-header>
+                    VulnerabilityTimeline
+                  </nb-card-header>
+                  <nb-card-body  >
+                    <p-timeline *ngIf="events.length>0" [value]="events">
+                      <ng-template pTemplate="content" let-event>
+                        <small class="p-text-secondary">{{ event.date }}</small>
+                      </ng-template>
+                      <ng-template pTemplate="opposite" let-event>
+                        {{ event.status }}
+                      </ng-template>
+                      <ng-template pTemplate="marker" let-event>
+                        <nb-icon  [icon]="event.icon" pack="eva" [style.backgroundColor]="event.color" style="border-radius: 7px"></nb-icon>
+                      </ng-template>
+                    </p-timeline>
+                  </nb-card-body>
+                </nb-card>
+              </div>
+            </div>
             <div class="row">
               <div class="col-md-12">
                 <nb-card>
@@ -130,8 +162,35 @@ export class DetailsComponent implements ViewCell, OnInit {
   references: string;
   grade: number;
   bugTrackerEnabled: boolean;
+  vulnAudits: ProjectAudit[];
+  vuln: VulnerabilityHistoryRequest = new VulnerabilityHistoryRequest();
+  events: EventItem[] = [];
   constructor(private windowService: NbWindowService, private showProjectService: ShowProjectService,
               private toast: Toast) {
+  }
+  loadVulnAudit(source, id) {
+    return this.showProjectService.getHistoryForVuln(id, this.vuln, source).subscribe(data => {
+      this.events.push({ status: 'Asset Created', date: '', icon: 'edit-outline', color: 'rgba(102,243,232,0.71)' });
+      for (const item of data) {
+        if (item.eventType === 'CREATED') {
+          this.events.push(
+            {status: 'Vulnerability Added', date: item.start + ', number: ' + item.occurances, icon: 'plus-outline',
+              color: 'rgba(218,58,89,0.71)'},
+          );
+        } else if (item.eventType === 'UPDATED') {
+          this.events.push(
+            {status: 'Vuln re-detected', date: item.start + ', number: ' + item.occurances, icon: 'share-outline',
+              color: 'rgba(41,110,185,0.71)'},
+          );
+
+        } else if (item.eventType === 'RESOLVED') {
+          this.events.push(
+            {status: 'Resolved', date: item.end + ', number: ' + item.occurances, icon: 'checkmark-outline',
+              color: 'rgba(96,164,75,0.71)'},
+          );
+        }
+      }
+    });
   }
 
   loadVulnerability() {
@@ -139,13 +198,26 @@ export class DetailsComponent implements ViewCell, OnInit {
       this.vulnerability = data;
       this.location = data.location;
       this.inserted = data.inserted;
+      this.vuln.location = data.location;
+      this.vuln.vulnerability = this.vulnerability.vulnerability.name;
+
       if (data.vulnerabilitySource.name === 'Network') {
+        this.loadVulnAudit('interface', data.anInterface.id);
         this.asset = data.anInterface.asset.name + ' / ' + data.anInterface.privateip;
       } else if (data.vulnerabilitySource.name === 'SourceCode') {
+        this.loadVulnAudit('code', data.codeProject.id);
+        this.asset = data.codeProject.name;
+      } else if (data.vulnerabilitySource.name === 'GitLeaks') {
+        this.loadVulnAudit('code', data.codeProject.id);
+        this.asset = data.codeProject.name;
+      } else if (data.vulnerabilitySource.name === 'IaC') {
+        this.loadVulnAudit('code', data.codeProject.id);
         this.asset = data.codeProject.name;
       } else if (data.vulnerabilitySource.name === 'WebApplication') {
+        this.loadVulnAudit('webapp', data.webApp.id);
         this.asset = data.webApp.url;
       } else if (data.vulnerabilitySource.name === 'OpenSource') {
+        this.loadVulnAudit('code', data.codeProject.id);
         this.asset =  data.codeProject.name;
       } else if (data.vulnerabilitySource.name === 'CISBenchmark') {
         this.asset = data.location;
@@ -159,6 +231,7 @@ export class DetailsComponent implements ViewCell, OnInit {
         { title: this.vulnerability.vulnerability ? this.vulnerability.vulnerability.name :
             this.vulnerability.cisRequirement.name, hasBackdrop: true },
       );
+
     });
   }
   ngOnInit(): void {