+++ src/app/pages/show-project/show-project-details-chart/vuln-trend-pie.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnDestroy } from '@angular/core';
+import {AfterViewInit, Component, Input, OnChanges, OnDestroy, OnInit, SimpleChanges} from '@angular/core';
 import { NbThemeService } from '@nebular/theme';
 import {ShowProjectService} from '../../../@core/service/ShowProjectService';
 import {ActivatedRoute, Router} from '@angular/router';
 import {Severities} from '../../../@core/Model/Severities';
 import {ProjectConstants} from '../../../@core/constants/ProjectConstants';
+import { ChangeDetectorRef } from '@angular/core';
 
 @Component({
   selector: 'ngx-vuln-trend-pie',
@@ -11,103 +12,117 @@ import {ProjectConstants} from '../../../@core/constants/ProjectConstants';
     <div echarts [options]="options" class="echart"></div>
   `,
 })
-export class VulnTrendPieComponent implements OnDestroy {
+export class VulnTrendPieComponent implements OnDestroy, OnInit, AfterViewInit, OnChanges {
   options: any = {};
   themeSubscription: any;
   _entityId: number;
-  severities: Severities;
   constants: ProjectConstants = new ProjectConstants();
+  @Input() severities: Severities;
 
   constructor(private theme: NbThemeService, private showProjectService: ShowProjectService,
-              private _route: ActivatedRoute, private router: Router) {
+              private _route: ActivatedRoute, private router: Router, private cdr: ChangeDetectorRef) {
     this._entityId = +this._route.snapshot.paramMap.get('projectid');
     if (!this._entityId) {
       this.router.navigate(['/pages/dashboard']);
     }
-    this.loadSeveritiesChart();
+
   }
-  loadSeveritiesChart() {
-    return this.showProjectService.getSeverityChart(this._entityId).subscribe(data => {
-      this.severities = data;
-      this.drawChart();
-    });
+  ngOnChanges(changes: SimpleChanges) {
+    this.options = {
+      xAxis: {
+        type: 'category',
+        data: ['Critical', 'High', 'Medium', 'Low'],
+      },
+      yAxis: {
+        type: 'value',
+      },
+      series: [
+        {
+          data: [
+            { value: this.severities?.Critical, itemStyle: { color: '#F74646' } },
+            { value: this.severities?.High, itemStyle: { color: '#EBC04A' } },
+            { value: this.severities?.Medium, itemStyle: { color: '#A4EEDA' } },
+            { value: this.severities?.Low, itemStyle: { color: '#9FFAF1' } },
+
+          ],
+          type: 'bar',
+        },
+      ],
+    };
   }
 
+  ngOnInit() {
+  }
+  ngAfterViewInit() {
+  }
   drawChart() {
-    this.themeSubscription = this.theme.getJsTheme().subscribe(config => {
-      const colors = config.variables;
-      const echarts: any = config.variables.echarts;
-
-      this.options = {
-        backgroundColor: echarts.bg,
-        color: [ colors.successLight, colors.primaryLight, colors.warningLight, colors.dangerLight ].reverse(),
-        tooltip: {
-          trigger: 'item',
-          formatter: '{a} <br/>{b} : {c} ({d}%)',
-        },
-        legend: {
-          orient: 'vertical',
-          left: 'left',
-          data: ['Low', 'Medium', 'High', 'Critic'],
-          textStyle: {
-            color: echarts.textColor,
+    if (this.severities) {
+      this.themeSubscription = this.theme.getJsTheme().subscribe(config => {
+        const colors = config.variables;
+        const echarts: any = config.variables.echarts;
+        this.options = {
+          tooltip: {
+            trigger: 'item',
           },
-        },
-        series: [
-          {
-            name: this.constants.PROJECT_CHARTS_SEVERITY_TITLE,
-            type: 'pie',
-            radius: ['40%', '70%'],
-            center: ['50%', '50%'],
-            avoidLabelOverlap: true,
-            data: [
-              {
-                'name': 'Critic',
-                'value': this.severities.Critical,
+          legend: {
+            // Try 'horizontal'
+            orient: 'vertical',
+            left: 10,
+            top: 'center',
+          },
+          color: [colors.successLight, colors.primaryLight, colors.warningLight, colors.dangerLight].reverse(),
+          series: [
+            {
+              name: 'Access From',
+              type: 'pie',
+              radius: ['40%', '90%'],
+              avoidLabelOverlap: false,
+              itemStyle: {
+                borderRadius: 10,
+                borderColor: '#fff',
+                borderWidth: 2,
               },
-              {
-                'name': 'High',
-                'value': this.severities.High,
+              label: {
+                show: false,
+                position: 'center',
               },
-              {
-                'name': 'Medium',
-                'value': this.severities.Medium,
-              },
-              {
-                'name': 'Low',
-                'value': this.severities.Low,
-              },
-            ],
-            itemStyle: {
               emphasis: {
-                shadowBlur: 10,
-                shadowOffsetX: 0,
-                shadowColor: echarts.itemHoverShadowColor,
-              },
-              borderRadius: 10,
-              borderColor: '#fff',
-              borderWidth: 2,
-            },
-            label: {
-              normal: {
-                textStyle: {
-                  color: echarts.textColor,
+                label: {
+                  show: true,
+                  fontSize: 40,
+                  fontWeight: 'bold',
                 },
               },
-            },
-            labelLine: {
-              normal: {
-                lineStyle: {
-                  color: echarts.axisLineColor,
-                },
+              labelLine: {
+                show: false,
               },
+              data: [
+                {
+                  'name': 'Critic',
+                  'value': this.severities?.Critical,
+                },
+                {
+                  'name': 'High',
+                  'value': this.severities?.High,
+                },
+                {
+                  'name': 'Medium',
+                  'value': this.severities?.Medium,
+                },
+                {
+                  'name': 'Low',
+                  'value': this.severities?.Low,
+                },
+              ],
             },
-          },
-        ],
-      };
-    });
+          ],
+        };
+      });
+    }
   }
 
   ngOnDestroy(): void {
+    this.themeSubscription.unsubscribe();
   }
+
 }